requirements.txt
flask>=3.0.0
python-dotenv>=1.0.1
openai>=1.45.0
langchain>=0.2.14
langchain-community>=0.2.12
langchain-openai>=0.1.22
pydantic>=2.6.0
duckduckgo-search>=5.3.1
tiktoken>=0.7.0
gunicorn>=21.2.0
